+++ libft/ft_strlcpy.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jhii <jhii@student.42kl.edu.my>            +#+  +:+       +#+        */
+/*   By: jhii <jhii@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/08 22:13:19 by jhii              #+#    #+#             */
-/*   Updated: 2021/07/21 04:37:35 by jhii             ###   ########.fr       */
+/*   Updated: 2022/05/18 16:08:08 by jhii             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlcpy(char *dst, const char *src, size_t size)
+int	ft_strlcpy(char *dst, const char *src, unsigned long size)
 {
-	size_t	i;
+	int	i;
 
 	if (!dst || !src)
 		return (0);
